+++ typescript/models/Role.ts
@@ -1,7 +1,7 @@
 import {Column, DataType, Model, Table} from 'sequelize-typescript';
 
 @Table({tableName: 'roles', timestamps: false})
-export default class Roles extends Model<Roles> {
+export default class Role extends Model<Role> {
     @Column({type: DataType.INTEGER, primaryKey: true, autoIncrement: true})
       id: number;
